Clean up, IDisposable, Fixed References
Removed unnecessary References from Wetr.Generator (Mahapss, ControlEZ)

Implemented IDisposable for SimulationViewModel since we use timers

Added GetHashCode() for all Domain Classes that have the Equals function. Ist that important you ask? ->
Yes, it is important if your item will be used as a key in a dictionary, or HashSet<T>, etc - since this is used (in the absence of a custom IEqualityComparer<T>) to group items into buckets. If the hash-code for two items does not match, they may never be considered equal (Equals will simply never be called).